Rename zoom-in to zoom plugin
