+++ packages/rpv-zoom/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@phuocng/rpv-zoom-in",
+    "name": "@phuocng/rpv-zoom",
     "version": "2.0.0",
     "description": "A React component to view a PDF document",
     "license": "https://react-pdf-viewer.dev/license",
@@ -19,7 +19,7 @@
     "repository": {
         "type": "git",
         "url": "https://github.com/phuoc-ng/react-pdf-viewer",
-        "directory": "packages/rpv-zoom-in"
+        "directory": "packages/rpv-zoom"
     },
     "bugs": {
         "url": "https://github.com/phuoc-ng/react-pdf-viewer/issues"
