+++ packages/rpv-zoom/src/StoreProps.ts
